tests: Migrate to DvtProvider
